+++ ChangeLog.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.8] -2019-12-28
+
+### Added
+
+- Add focal loss on mask
+
+### Fixed
+
+- Heatmap now works well
+
 ## [0.0.7] -2019-12-17
 
 ### Added
@@ -13,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Heatmap of mask
 - Now we can decide if to use optimization of prediction (shorten time)
 - Another base model: wide-resnet-101
+- Use mask dataset information
 
 ### Changed